+++ level_editor/src/org/vertexarmy/dsr/leveleditor/editors/polygon/PolygonEditorListener.java
@@ -1,0 +1,9 @@
+package org.vertexarmy.dsr.leveleditor.editors.polygon;
+
+/**
+ * created by Alex
+ * on 08-Apr-2015.
+ */
+public interface PolygonEditorListener {
+    void deletePolygonRequested();
+}